notification: Fix ::handle-* return values
When handling incoming method calls via the generated ::handle-*
signals, handlers are expected to return %TRUE to indicate that
the invocation has been handled. In the case that no handler handles
the invocation, the method is considered unknown and a corresponding
D-Bus error is returned.

As we currently connect handlers with no return value at all, we end
up with a Schroedinger method that may or may not exist - as the gtk
code path calls methods that do return %TRUE, there's a good chance
that we are lucky and something true-ish is picked from the stack,
but for the fdo path all bets are off.

Fix this by moving the return values to the actual signal handlers.