notification: Fix 'icon' lookup
The type of a serialized GIcon is either 's' or '(sv)', not 'v'.
We could use one type first and retry with the second one, but given
that xdg-desktop-portal already validates the input, we can just
accept any type and rely on icon deserialization to catch unlikely
errors (say: calling the method manually via d-feet).