notification: Support raw data in 'image-data' hint
In addition to file- and icon theme name, the freedesktop
notification spec also allows to specify images as raw data,
using its own serialization protocol. It's not pretty, but
libnotify provides a convenient API for it with the result
that it's used more than it should be, so support it in the
portal as well.