screencastwidget: Keep internal selection state updated
Currently, interacting with the screencast dialog with both
keyboard and cursor is problematic due to how it internally
keeps selection accounted.

The bulk of the problem is that the current code tracks selected
rows internally by setting custom data on them. This custom
data is set only when activating rows, but not when navigating
using the keyboard.

This is visible by:

 1. Activating any row
 2. Moving up or down with the keyboard
 3. Activating the same row again

Fix that by keeping the internal selection state updated in the
idle callback.