Minor fixes to screensharing (#214)
* screencast: Remove unused variable

* screencastwidget: Emit HAS_SELECTION_CHANGED in idle

GtkListBox activates rows after selecting them, which causes
the signal to be emitted twice:

 1. By GtkListBox itself, in response to GtkGestureMultiPress
 2. By us in response to (un)selecting the row in the row-activated
    callback.

Move the signal emission code to an idle callback, and prevent
the HAS_SELECTION_CHANGED signal to be emitted twice. It will
also allow us to fix keyboard handling in the next commit.

* screencastwidget: Keep internal selection state updated

Currently, interacting with the screencast dialog with both
keyboard and cursor is problematic due to how it internally
keeps selection accounted.

The bulk of the problem is that the current code tracks selected
rows internally by setting custom data on them. This custom
data is set only when activating rows, but not when navigating
using the keyboard.

This is visible by:

 1. Activating any row
 2. Moving up or down with the keyboard
 3. Activating the same row again

Fix that by keeping the internal selection state updated in the
idle callback.