Add window screen cast
This adds window capture support to the screen cast recording and
dialog.

It works by retrieving the list of windows from GNOME Shell using
org.gnome.Shell.Introspect. Currently said API does not emit signals
when windows are added or removed, so it just retrieves the list when
the widget is shown.

Eventually it should be able to auto-update as windows are added and
removed by GNOME Shell.

[jadahl: Added commit message, made window list scrollable, added
ShellIntrospect state tracking object]