Ensure FileChooser::SaveFiles give URIs for non-existent files
Check that the required files don't already exist in the selected folder
before reurning a URI to them. If a file does already exist, append a
"(n)" from 1..n to the basename in front of the extension (if any) until
a name to a non-existing file is generated, and use that instead.