Implementation of org.freedesktop.portal.FileChooser::SaveFiles (#236)
* Implementation of org.freedesktop.portal.FileChooser::SaveFiles

GTK implementation of new `SaveFiles` method using a
`GTK_FILE_CHOOSER_ACTION_SELECT_FOLDER` file chooser.

See https://github.com/flatpak/xdg-desktop-portal/pull/368

* Ensure FileChooser::SaveFiles give URIs for non-existent files

Check that the required files don't already exist in the selected folder
before reurning a URI to them. If a file does already exist, append a
"(n)" from 1..n to the basename in front of the extension (if any) until
a name to a non-existing file is generated, and use that instead.