background: Emit RunningApplicationsChanged (#259)
* background: Emit RunningApplicationsChanged

Listen to the gnome-shell signal of the same name,
compare the app state we got, and emit our own signal
if any apps changed to or from background state.

* Add a background permission dialog

Just so we can pop up something when the notification is clicked.

* background: Use a different shell property

We were getting confused by the hidden state for
windows on other workspaces. Just use the running
applications that the shell reports.