Use the application name in the print notification
When printing with an flatpak application, the notification
telling that it prints says "Document from %s" where %s is the
app desktop file id (i.e. will print "Document from
org.gnome.Evince" with evince).

This is not very clear what this name is in the notification,
particularly with a third-party app, for example
com.github.my_beautiful_github_name.app_name.

So this changes %s to the actual translated application name,
looking in the corresponding desktop file entry.