wallpaper: Drop "lockscreen" option
In GNOME 3.36 the lockscreen is now a blurried version of the
wallpaper. Although the gsetting for setting the lockscreen
background image will persist available, we should be consistent
with GNOME's UX decisions.

The frontend code should keep the "set-on" option, since other
desktop environments such as KDE might still have a separate
lockscreen image.

The "set-on" option will be simply ignored for the GTK backend.

See https://blogs.gnome.org/shell-dev/2020/02/18/login-and-unlock-in-gnome-shell-3-36/