screencast: Force window icon size (#344)
Some applications do not provide SVG icons, nor a range of icons,
only 64x64 or bigger. In these cases, even if the GtkImage is
created with the GTK_ICON_SIZE_DND (32px) hint passed, the window
icons would be bigger.

Fix that by forcing a 32px pixel size to the GtkImage.