Don't leak app_name string if a desktop file or it's "Name" is absent
Note that earlier the fallback app_name may or may not have included
the .desktop extension as a suffix, depending on what was passed as
app_id to the function.  Now, it will always include the extension.

As a fallback, this shouldn't matter and being more predictable is
likely a good thing.

Fallout from 68c044c2c97261baa9b154ad29ab3ec826dd26b4