notification: Use notification desktop-id to select backend and notify
Notifications can provide a specific desktop-id, in such case we should
just use the fdo notifications (to ensure that such info is consumed by
the shell) if the provided ID is not a FDO application ID and pass such
value to it, while when it's a valid application ID we can just use the
gtk backend to have the full experience.

When desktop-id is not provided by the portal, it means it is matching
the application-id, so we consider them the same.