notification: Use always private GNOME Shell interface when available
We couldn't use the private GNOME Shell interface because it was missing
a couple of features we need in specific instances.
The private GNOME Shell interface can now also emit the ActionInvoked
signal and since version v2 it's mandatory that apps specify a desktop
file id so we can do the switch.