Update our CI pipeline (#518)
* ci: Use a custom container image

We should crib the CI setup from xdg-desktop-portal, which uses a
separate job to build a custom container image in which to run the whole
pipeline. This allows us to clone xdg-desktop-portal just once, instead
of every time, and gives us a more resilient pipeline.

* ci: Add release job

Build and publish a release directly from CI, like we're actual adults.

* Re-format the NEWS file

Put the release date on its own line, for future reference.

Do this only for the current minor version, to avoid excessive churn.