notification: Emit platform data on notification Activation
The action platform data can be provided by the shell using the FDO
notification protocol.

This information is currently parsed and used for FDO Application
activation, but not passed to the portal in case different
implementations are used.

Notification portal v2 however support sending platform data now as the
second parameter of the parameters array.

While the GTK portal is still using the notification protocol v1, the
portal v2 is an addition of the current protocol and so we can safely
just pass such information back and clients should have no problem with
it.

In case no action target is provided by the application we still need to
send an ordered array to the clients, so in such case send them back the
an array. Using an empty tuple or a nothing maybe does not seem to work
for this, but in any case it's just a place-holder since if we get there
it means that the application sent us nothing and so we've nothing to
emit back